+++ src/main/java/chess/controller/ChessController.java
@@ -3,8 +3,6 @@ package chess.controller;
 import chess.domain.ChessGame;
 import chess.domain.command.Command;
 import chess.domain.command.Commands;
-import chess.domain.dto.BoardDto;
-import chess.domain.dto.PointDto;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -17,20 +15,20 @@ public class ChessController {
             final Commands inputCmd = InputView.inputCommand();
             Command command = Command.of(inputCmd.mainCommand());
             command.apply(chessGame, inputCmd);
-            view(chessGame, command); // view 작업을 여기서 해줌.
+            view(chessGame, command);
         }
     }
 
     private void view(ChessGame chessGame, Command command) {
         if (command.isStart()) {
-            OutputView.printBoard(new BoardDto(chessGame.board(), chessGame.turn()));
+            OutputView.printBoard(chessGame.boardDto());
         }
         if (command.isMove()) {
-            OutputView.printBoard(new BoardDto(chessGame.board(), chessGame.turn()));
+            OutputView.printBoard(chessGame.boardDto());
             confirmKingDead(chessGame);
         }
         if (command.isStatus()) {
-            OutputView.printStatus(new PointDto(chessGame.calculatePoint()));
+            OutputView.printStatus(chessGame.pointDto());
         }
     }
 
@@ -39,33 +37,4 @@ public class ChessController {
             OutputView.printWinner(chessGame.winner());
         }
     }
-//    public static void start(ChessGame chessGame, Commands command) {
-//        chessGame.initBoard(BoardInitializer.init());
-//        OutputView.printBoard(new BoardDto(chessGame.board(), chessGame.turn()));
-//    }
-//    public static void move(ChessGame chessGame, Commands command) {
-//        if (chessGame.isReady() || chessGame.isEnd()) {
-//            throw new IllegalArgumentException("[ERROR] 게임이 초기화되지 않았거나, 종료되지 않았습니다.");
-//        }
-//        chessGame.move(command);
-//        OutputView.printBoard(new BoardDto(chessGame.board(), chessGame.turn()));
-//        confirmKingDead(chessGame);
-//    }
-//
-//    public static void end(ChessGame chessGame, Commands command) {
-//        chessGame.endGame();
-//    }
-//
-//    public static void status(ChessGame chessGame, Commands command) {
-//        if (chessGame.isReady()) {
-//            throw new IllegalArgumentException("[ERROR] 게임이 초기화되지 않았습니다.");
-//        }
-//        OutputView.printStatus(chessGame.calculatePoint());
-//    }
-//
-//    private static void confirmKingDead(ChessGame chessGame) {
-//        if (chessGame.isEnd()) {
-//            OutputView.printWinner(chessGame.winner());
-//        }
-//    }
 }
