+++ src/test/java/chess/domain/command/CommandTest.java
@@ -1,6 +1,5 @@
-package chess.domain;
+package chess.domain.command;
 
-import chess.domain.command.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
