+++ src/test/java/chess/domain/utils/BoardInitializerTest.java
@@ -1,6 +1,5 @@
-package chess.domain;
+package chess.domain.utils;
 
-import chess.domain.utils.BoardInitializer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
